dma-heap: Switch to OwnedFd
Rust 1.63 improved the handling of file descriptors by including the new
OwnedFd and BorrowedFd structures.

This clarifies the semantics of the file descriptors we can return, and
most importantly around who has to close the file descriptor.

We also had that ambiguity, where depending on the structure
implementing the FromRawFd trait the caller was using, we would require
or not the caller to explicitly close the file descriptor we returned.

This essentially meant that we would expect the caller responsible of
closing the FD we allocated, which is what OwnedFd is about. Make it
clear and unambiguous by switching to an OwnedFd.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>